Version 2.3.0
In the version 2.3.0 are many bug fixes included but also a small innovation.
The biggest bug that was fixed concerns the config. The config was not created correctly. This bug was finally fixed with this update.

But you have to rebuild the complete folder structure. I have said this many times before. But this time it is really the last time that you have to rebuild the config.
As a new feature titles have been added. For this you can find settings in Config.yml and Messages.yml.

Bug Fixes

- /wbs giveall -> As soon as a player has a full inventory the process is aborted. FIXED

- ItemNumbers -> If the ItemNumbers function is set to false, WonderBag 2 and WonderBag 3 cannot be bought, no matter if Chest or Item. Also affects /wbs give, /wbs gift, wbs giveall. FIXED

- If you want to empty your inventory with shift, this is done asynchronously. FIXED

- The sound NoInventorySpace is played although NoInventorySpace_Enable is set to false. FIXED

In version 2.4.0 the Settings GUI will be implemented completely.