BUGFIX Version 2.3.1
Fixed a bug that caused a NullPointerException to occur when placing a chest.

If you gave a chest the same name as the GUI, then the chest was recognized as a GUI. This has been fixed.

Added a new option to set the chest to be removed from the inventory when in the Creative gamemode.